[KARAF-2442] Add Role-based access to console/shell commands
This commit utilizes KARAF-2455 to add role-based access to console
commands.
Karaf commands are registered as services, and these services have ACLs
defined for role-based access. However, the level at which these ACLs
are defined is a little too fine grained to be user friendly for
commands. The service-level ACLs are defined per-service while the ACLs
for commands would be ideally defined per scope, e.g. all the ACLs for
all the commands relating to one scope in one file.
Therefore this commit contains a transformer that takes ACL definitions
for commands in a single scope and turns it into separate Config Admin
definitions for each associated service.
The command ACLs are defined using OSGi ConfigAdmin and must have a PID
that follows this structure: org.apache.karaf.command.acl.<scope>
Then command ACL definition has the following form:
  command1 = role1, role2
  command2 = role1, role2
Argument matching is also supported via
  command3[/.*regexp.*/] = role3
Note that all the arguments passed to the command are represented as a
single string (with spaces to separate).

When logging in remotely via SSH the roles associated with the current
user are obtained via JAAS.
When logging in locally via bin/karaf the roles defined in the
  karaf.local.roles
property (in system.properties) are used. The default roles are
  admin,manager,viewer