[KARAF-4145] KAR Mojo: fix metadata generation of artifacts
Problem description:
* use the kmp to generate a kar that contains snapshot version
* use the kmp to generate a custom distribution that fills the local
  repository of the distribution with the content of that kar
* try to install the snapshot artifact (not using any other repository
  then the system directory
* if your snapshots has been timestamped and a maven metadata XML does
  not exist on KAR creation the installation of the artifact will fail

The KAR Mojo contains a workaround to use the base version of a snapshot
instead of a timestamped one because it does not work in
startup.properties.
The metadata that is used for a snapshot artifact is generated (if not
present) using the timestamped version. Also if there exists already a
metadata, we can not use them without further checking for the
version...
Let's move the version fix in front of the metadata generation and
generate the metadata for every snapshot.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>