Do also consider transitive dependencies when using version ranges
Sometimes is also necessary that transitive dependencies are declared
with a version range in the resulting feature file. This can be achieved
with includeTransitiveVersionRanges=true. When set, the plugin will
search the dependency in the model-graph and will use the same version
range as declared in the POM in the resulting feature file.