KARAF-4897 - Add an option to simplify bundle dependencies
This adds a simplifyBundleDependencies option to
GenerateDescriptorMojo, disabled by default, which drops bundle
dependencies from features if the bundles are included in "parent"
features. I.e. if feature A depends on bundle B and feature C, and
feature C also includes bundle B, then feature A will be reduced to
only depend on feature C, relying on the latter to supply bundle B.

Signed-off-by: Stephen Kitt <skitt@redhat.com>