Remove the VM option "-server"
If not explicitly defined, the vm option `-server` has been 
automatically set since Java 5.  The bar for it picking `-server` over 
`-client` is pretty low by modern standards; 2 CPU and 2GB RAM.

http://docs.oracle.com/javase/8/docs/technotes/guides/vm/server-class.html

The Oracle ARMv7 HFP does not support `-server`.  So when running on a 
Raspberry Pi, one must either set JAVA_OPTS or regex replace the 
relevant lines in launch scripts.

IMO it would be better to remove this _redundant_ flag, as I'd guess 
there are few systems that would actually require this to be explicitly 
set.