[KARAF-5414] Handle varying feature dependency versions
When a feature stub specifies a feature dependency, and the POM
contains a slight variant of the dependency (e.g., the former
specifies a version range, the latter specifies a single version), we
currently ignore the latter when resolving the former. This patch
changes the behaviour so that feature dependencies are matched more
generously, which allows dependency information to be correctly
analysed.

Signed-off-by: Stephen Kitt <skitt@redhat.com>