Fix meta command / completer to not print a warning when MetaType is not available
I think this is a regression caused by https://github.com/apache/karaf/commit/bf666631fe84f0394a9b948b5088d205bf52db36#diff-290cb1e2332aeecaf61aaa6aa3602768
which removes the anonymous inner classes and thus cause a CNFE to be thrown when the MetaCompleter getMethods() is called for introspection.
So this commit slightly change the mechanism and a warning is displayed when the completer is initialized, while a message is printed on the console when calling the command.