[KARAF-6125] Standardized timezone handling for audit dates
Fixed a couple of cases where the FastDateFormat was not using a consistent timezone (sometimes provided and sometimes the default). This also required fixing a couple of callers of the FastDateFormat, which was mostly related to tests.
This bug caused audit file names to sometimes incorrectly roll between days depending on the timezone of the user.