KARAF-6207, bootFeatures sometimes being ignored
According to the documentation, the maven-karaf-plugin will install
features listed in the installedFeatures in the "system" internal
repository. Features listed in bootFeatures are additionally added to
the boot-features in the features service configuration file.

Hence, the configuration options do not conflict and a configuration
like this should result in the features being installed and added to the
boot-features configuration:

    <configuration>
      <bootFeatures>
        <feature>foo</feature>
      </bootFeatures>
      <installedFeatures>
        <feature>foo</feature>
      </installedFeatures>
    </configuration>

In fact, the configuration should be equivalent to `foo` being listed
just as a bootFeature.

But what actually happens is that the installedFeature will overwrite
the bootFeature and the feature is not added to the boot-features
configuration.

This patch switches the handling of bootFeatures and installedFeatures
which makes a feature configured in both sections a bootFeature,
adhering to the specifications of both configurations.