[KARAF-6274] Karaf 4.2.5 maven plugin breaks if no archives are generated
Karaf 4.2.5 includes a line checking if archiveTarGz or archiveZip is
selected, throwing an error otherwise. Unfortunately, this line breaks
our build and a few of our use cases where we do not pack our
assemblies.

For example, we have a development assembly which we have end up in a
special folder unpacked and prepared to be launched. We could, of
course, pack and then automatically unpack it but that would just be a
waste of time.

We also have some assemblies which we pack with additional files
(documentation, …) which is handled separately. Again a use-case for
assemblies unpacked by the karaf-maven-plugin.

The change in question is commit 7da5044 (pull request #811)