[KARAF-6351] Handle package-less classes
The classes command unconditionnally tries to extract a package name,
which fails with classes which don't have one. This patch checks for
the presence of "/" before attempting to use it as a delimiter.
Classes with no package are assumed to not be exported.

Signed-off-by: Stephen Kitt <skitt@redhat.com>

(cherry-pick from commit 5e885946aebd43c01bd816b3c71a85d1602f8362)