Put JAVA_OPTS between quotes for handling parameters with spaces
Certain Java options might contain spaces, for example the `-XX:OnError` which can be used to execute a command or script and takes the parameter `%p` as placeholder for the current PID of the Java process. However, due to the way e.g. Bash handles variables in scripts, it makes a lot of difference whether these are placed between quotes. Without quotes, spaces are handled so that they break up parameters.
This causes e.g. the following `export JAVA_OPTS="-XX:OnError=\"echo %p\""` to be handled as two separate arguments resulting in the error: `Error: Could not find or load main class %p`

The only downside I found from quoting is if e.g. a user has a script `export JAVA_OPTS="${JAVA_OPTS} ${MY_OPTS}"` and `JAVA_OPTS` was not previously set, an additional space is introduced which is not handled nicely.