Optimize docker script and update README.
Fixed typo error on exit code when checked that buildx is not installed

Brace expansions and globs are literal in assignments.

This \= will be a regular '=' in this context.

Double quote to prevent globbing and word splitting.

Use $(...) notation instead of legacy backticked `...`.

Change exit code to "2" because "1" already used

Remove double quote, are not necessary.

Updated documentation on using the build command (build.sh):
Direct docker commands have been removed from the documentation because
they are cumbersome to use due to the arguments to be specified.
Also the specified docker commands didn't work. For example, the command
`docker build -t karaf .` it can't work because the Dockerfile expects the
build parameter `karaf_dist`.