Fix version parsing
Make the first group of the regexp non-greedy to ensure it won't swallow
the build meta data part. In particular, with a greedy first group it
swallows everything up to the last dash that's followed by a
digit, something that's quite common in SNAPSHOT builds.
As a result a name like `foobar-1.0.0-PR-4-SNAPSHOT.jar` was parsed into
version `4.0.0-SNAPSHOT`.