KARAF-7289 - Provide config property name completion using MetaType info
ConfigurationPropertyCompleter now uses the MetaType service to retrieve
configuration property names in addition to the ones already defined in
the configuration.

Adds an utility method in MetaServiceCaller to find MetaType info for a
specified PID. This method accepts a function to do something with the
result. MetaCommand is refactored to use this function.

Change-Id: Ieab22ed2daaefa55675fbb2eb41e77ac534b5251