Add null guard to MavenDownloadTask
`resolver` is a private property of the class `MavenDownloadTask` that
has no corresponding setter method.
As such, if `null` is provided as the `resolver` to the constructor, it
cannot be changed elsewhere and is bound to throw a NullPointerException
at a later time.
For debugging purposes, however, a fail-fast guard is preferable, which
this commit implements.